requirements.txt
requests==2.31.0
msal==1.24.1
python-dotenv==1.0.0
streamlit==1.28.1
flask==2.3.3
flask-cors==4.0.0 